database connection
insert and select

+++ CSharp_Assignment4/CSharp_Assignment4/Product_Detail.cs
@@ -12,7 +12,7 @@ namespace CSharp_Assignment4
 {
     public partial class Product_Detail : Form
     {
-        public Product_Detail(string Name, int Inventory, int Count, Double Price, DateTime Date, int Number)
+        public Product_Detail(string Name, int Inventory, int Count, int Price, DateTime Date, int Number)
         {
             InitializeComponent();
             lbl_Name.Text = "Name: " + Name;
